Changes: - Sound mixer and synth fix by @jpfli. - Refactored to use common audio interrupt implementation for HW and SIM.
