+++ Pokitto/POKITTO_SIM/SimSound.cpp
@@ -101,155 +101,8 @@ void pokPlayStream() {
     streamon=1;
 }
 
-void pokSoundIRQ() {
-    #ifndef POK_SIM
-    pwmout_t* obj = &audiopwm;
-    #endif
-    #if POK_STREAMING_MUSIC > 0
-        #if POK_STREAMFREQ_HALVE
-        streamstep = 1-streamstep;
-        #else
-        streamstep = 1;
-        #endif // POK_STREAMFREQ_HALVE
-        streamstep &= streamon; // streamon is used to toggle SD music streaming on and off
-        if (streamstep) {
-            // Get byte from the stream buffer
-            output = (*currentPtr++);
-
-            // If exists, mix the sound effect to the output.
-            if( Pokitto::Sound::sfxDataPtr != Pokitto::Sound::sfxEndPtr ){
-
-                // Get sfx byte
-                uint8_t sfxSample = 0;
-                if( Pokitto::Sound::sfxIs4bitSamples ) {
-                    if(Pokitto::Sound::sfxBytePos++ == 0) {
-                        sfxSample = (*Pokitto::Sound::sfxDataPtr) & 0xf0;  // 4-bit sample is in the high nibble
-                    }
-                    else
-                    {
-                        sfxSample = *Pokitto::Sound::sfxDataPtr << 4;  // 4-bit sample is in the low nibble
-                        Pokitto::Sound::sfxDataPtr++;
-                        Pokitto::Sound::sfxBytePos = 0;
-                    }
-                }
-                else {
-                    sfxSample = (*Pokitto::Sound::sfxDataPtr++);  // 8-bit sample
-                }
-
-                #ifdef PROJ_DISABLE_MIXING_SFX_WITH_SD_STREAMING
-                int32_t s = int32_t(sfxSample) - 1;
-                #else
-                int32_t s = (int32_t(output) + int32_t(sfxSample)) - 128;
-                #endif
-
-                if( s < 0 ) s = 0;
-                else if( s > 255 ) s = 255;
-
-                // streaming and sfx result
-                output = s;
-            }
-
-            // // streaming and sfx result
-            if(streamvol) {
-                output >>= 3-streamvol;
-                streambyte = output;
-            } else {
-                streambyte = 0; // duty cycle
-                output = 0;
-            }
-            if (currentPtr >= endPtr)
-            {
-            currentBuffer++;
-            if (currentBuffer==4) currentBuffer=0;
-            currentPtr = buffers[currentBuffer];
-            endPtr = currentPtr + BUFFER_SIZE;
-            }
-        }
-    #endif // POK_STREAMING_MUSIC
-
-    /** DO ADDITIONAL SOUND PROCESSING (NOT STREAM) OF SOUND HERE **/
-
-    #if POK_ENABLE_SYNTH
-        /** if song is being played from sd **/
-        if (playing) {
-                notetick++;
-                updatePlayback();
-                //updatePlaybackSD(playerpos&7);
-        }
-        /** oscillators update **/
-        osc1.count += osc1.cinc + (osc1.pitchbend); // counts to 65535 and overflows to zero WAS 8 !
-        osc2.count += osc2.cinc + (osc2.pitchbend); // counts to 65535 and overflows to zero
-        osc3.count += osc3.cinc + (osc3.pitchbend); // counts to 65535 and overflows to zero
-        #if POK_ALT_MIXING > 0 // heaviest cpu load, recalculate envelopes on each cycle
-        Marr[3]();
-        Marr[2]();
-        Marr[1]();
-        if (tick==0) Marr[0]();
-        Marr[0]();
-        #else
-        Marr[tick](); // call mixing function
-        #endif // ALT_MIXING
-        --tick;
-
-        /** mixing oscillator output **/
-        #ifdef POK_SIM
-        uint16_t op = (uint16_t) ((osc1.output)*(osc1.vol>>8))>>9;// >> 2 osc1.vol Marr;
-        op += (uint16_t) ((osc2.output)*(osc2.vol>>8))>>9;// >> 2 osc1.vol Marr;
-        op += (uint16_t) ((osc3.output)*(osc3.vol>>8))>>9;// >> 2 osc1.vol Marr;
-        #else
-        uint16_t op = (uint16_t) ((osc1.output)*(osc1.vol>>8))>>9;// >> 2 osc1.vol Marr;
-        op += (uint16_t) ((osc2.output)*(osc2.vol>>8))>>9;// >> 2 osc1.vol Marr;
-        op += (uint16_t) ((osc3.output)*(osc3.vol>>8))>>9;// >> 2 osc1.vol Marr;
-        #endif //POK_SIM
-
-        // Synth output
-        output = (uint8_t) op;
-
-    #endif // POK_ENABLE_SYNTH
-
-    #ifndef POK_SIM
-    /** HARDWARE **/
-        #if POK_ENABLE_SOUND
-            #if POK_STREAMING_MUSIC > 0
-                #if POK_STREAM_TO_DAC > 0
-                    #if POK_USE_DAC > 0
-                        if (streamstep) dac_write((uint8_t)streambyte); // duty cycle
-                    #endif // POK_USE_DAC
-                #else
-                    if (streamstep) {
-                            uint32_t t_on = (uint32_t)(((obj->pwm->MATCHREL0)*streambyte)>>8); //cut out float
-                            obj->pwm->MATCHREL1 = t_on;
-                    }
-                #endif // POK_STREAM_TO_DAC
-            #endif // POK_STREAMING_MUSIC
-
-            #if POK_STREAM_TO_DAC == 0
-                #if POK_USE_DAC > 0
-                    dac_write((uint8_t)output); // write synth output to DAC
-                #endif
-            #else
-                uint32_t t_on = (uint32_t)(((obj->pwm->MATCHREL0)*output)>>8); //cut out float
-                obj->pwm->MATCHREL1 = t_on;
-            #endif // POK_SYNTH_TO_DAC
-            soundbuf[soundbufindex++]=output;
-            if (soundbufindex==SBUFSIZE) soundbufindex=0;
-        #endif //POK_ENABLE_SOUND
-    #endif // HARDWARE
-
-    #ifdef POK_SIM
-    /** SIMULATOR **/
-        #if POK_STREAMING_MUSIC
-            //if (streamstep) {
-                uint32_t o = output + streambyte;
-                output = o/2;
-            //}
-        #endif // STREAMING
-        soundbyte = output;
-        soundbuf[soundbufindex++]=soundbyte/3;
-        if (soundbufindex==SBUFSIZE) soundbufindex=0;
-    #endif // POK_SIM
-}
-
+// *** INCLUDE THE COMMON IMPLEMENTATION FOR HWSound.cpp and SimSound.cpp.
+#include "../POKITTO_CORE/SoundCommonImp.h"
 
 /** THE FAKE ISR **/
